requirements.txt
biopython
click
edlib
joblib
numpy
matplotlib
pandas
plotly
python-slugify